Extract work tags panel
